Making test fixtures and memoized vals
Split tests across new classes
